Add fallback to 'charge_*' if 'energy_*' doesn't exist
On some laptops the 'energy_now' and 'energy_full' files don't exist,
but the 'charge_now' and 'charge_full' files exist instead, and with
'voltage_min_design' you can calculate the energy yourself.

For laptops where the 'energy_*' files exist, everything still works the
same as before, but if not, it tries to fallback to the 'charge_*' files
if available, before failing.